% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom Functions.R
\name{suggested_weight}
\alias{suggested_weight}
\title{Retrieve Suggested Weight for a Given CPS Variable}
\usage{
suggested_weight(variable, year = "2023")
}
\arguments{
\item{variable}{Character. The census variable for which the suggested weight is to be fetched.}

\item{year}{Character. The year for which to fetch the data. Default is "2023". Weights don't tend to change over the years.}
}
\value{
If the variable has a suggested weight, it prints the weight. If not, a message is shown indicating that the variable does not have a suggested weight.
}
\description{
This function fetches the suggested weight for a given CPS variable
from the U.S. Census Bureau's API for a specified year.
}
\examples{
suggested_weight("PEMLR")

}
\seealso{
\url{https://api.census.gov/data.html}
}
